evals/requirements.txt
frechet_audio_distance
resampy
laion_clap
torchlibrosa
ftfy
encodec
braceexpand
webdataset
wget
torchaudio
numpy
tqdm
mir_eval
